+++ src/light/shadows.c
@@ -3,15 +3,16 @@
 /*                                                        :::      ::::::::   */
 /*   shadows.c                                          :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: cle-tron <cle-tron@student.42barcelon      +#+  +:+       +#+        */
+/*   By: wlin <wlin@student.42.fr>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/12/24 15:19:29 by cle-tron          #+#    #+#             */
-/*   Updated: 2024/12/25 13:48:15 by cle-tron         ###   ########.fr       */
+/*   Updated: 2024/12/25 23:59:41 by wlin             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
+
 #include "miniRT.h"
 
-bool	closest_object(double intersection_distance, double min_distance)
+bool	closest_object(float intersection_distance, float min_distance)
 {
 	if (intersection_distance < min_distance)
 	{
@@ -21,9 +22,9 @@ bool	closest_object(double intersection_distance, double min_distance)
 	return (false);
 }
 
-t_ray	generate_light_ray(t_hit_rec *rec, double xyz_light[3])
+t_ray	generate_light_ray(t_hit_rec *rec, float xyz_light[3])
 {
-	double		scale[3];
+	float		scale[3];
 	t_ray		ray;
 
 	vec_scale(scale, rec->normal, 0.001);
@@ -33,9 +34,9 @@ t_ray	generate_light_ray(t_hit_rec *rec, double xyz_light[3])
 	return (ray);
 }
 
-double	calc_light_distance(double xyz_light[3], double ray_origin[3])
+float	calc_light_distance(float xyz_light[3], float ray_origin[3])
 {
-	double		tmp_vc[3];
+	float		tmp_vc[3];
 
 	vec_sub(tmp_vc, xyz_light, ray_origin);
 	return (vec_length(tmp_vc));
@@ -47,7 +48,7 @@ bool	calc_shadows(t_data *data, t_hit_rec *rec)
 	int			hit_flag;
 	t_obj		*obj;
 	t_hit_rec	tmp_rec;
-	double		min_distance;
+	float		min_distance;
 
 	obj = data->obj;
 	ray = generate_light_ray(rec, data->light->xyz);
